requirements.txt.txt
streamlit==1.28.0
Pillow==10.0.0
transformers==4.35.0
torch==2.1.0
opencv-python-headless==4.8.1